requirement.txt
streamlit
numpy
librosa
scikit-learn
plotly
Pillow
streamlit-webrtc
av
soundfile
reportlab
matplotlib